修改类型

+++ io.h
@@ -16,7 +16,7 @@ int writer_new_size(int fd, int size, writer_t **p_writer);
 
 int writer_delete(writer_t *writer);
 
-int writer_write(writer_t *writer, char *buf, int size);
+int writer_write(writer_t *writer, char *buf, size_t size);
 
 int writer_flush(writer_t *writer);
